Conform proto acquirelock requests to all proto requests
