+++ internal/app/subsystems/api/grpc/api/lock.proto
@@ -7,40 +7,42 @@ option go_package = "github.com/resonatehq/resonate/internal/app/subsystems/api/
 service Locks {
     rpc AcquireLock (AcquireLockRequest) returns (AcquireLockResponse) {}
     rpc HeartbeatLocks (HeartbeatLocksRequest) returns (HeartbeatLocksResponse) {}
-    rpc ReleaseLock (ReleaseLockRequest) returns (ReleaseLockResponse) {}  
+    rpc ReleaseLock (ReleaseLockRequest) returns (ReleaseLockResponse) {}
 }
-  
+
 message Lock {
     string resource_id = 1;
-    string process_id = 2; 
+    string process_id = 2;
     string execution_id = 3;
-    int64 expiry_in_milliseconds = 4;  
+    int64 expiry_in_milliseconds = 4;
     int64 expires_at = 5;
 }
 
 message AcquireLockRequest {
-    Lock lock = 1;
-    string request_id = 2;
+    string resource_id = 1;
+    string process_id = 2;
+    string execution_id = 3;
+    int64 expiry_in_milliseconds = 4;
+    string request_id = 5;
 }
-  
+
 message AcquireLockResponse {
     Lock lock = 1;
 }
-  
+
 message HeartbeatLocksRequest {
     string process_id = 1;
     string request_id = 2;
 }
-  
-message HeartbeatLocksResponse { 
+
+message HeartbeatLocksResponse {
     int32 locksAffected = 1;
 }
-  
+
 message ReleaseLockRequest {
     string resource_id = 1;
-    string execution_id = 2;   
+    string execution_id = 2;
     string request_id = 3;
 }
-  
+
 message ReleaseLockResponse {}
-
